Changed folio-number-input to kebab case
