+++ docs/iframe.html
@@ -130,4 +130,4 @@
             
         
             
-              window['FRAMEWORK_OPTIONS'] = {};</script><script src="runtime~main.7e8a6d162cd2179c37a9.bundle.js"></script><script src="vendors~main.7e8a6d162cd2179c37a9.bundle.js"></script><script src="main.7e8a6d162cd2179c37a9.bundle.js"></script></body></html>
+              window['FRAMEWORK_OPTIONS'] = {};</script><script src="runtime~main.1a03cf1521bd262149ef.bundle.js"></script><script src="vendors~main.1a03cf1521bd262149ef.bundle.js"></script><script src="main.1a03cf1521bd262149ef.bundle.js"></script></body></html>
