--- src/components/FolioNumberInput/index.ts
@@ -1,1 +1,0 @@
-export { default as FolioNumberInput } from './FolioNumberInput.vue'
