--- src/components/FolioNumberInput/package.json
@@ -1,11 +1,0 @@
-{
-    "name": "@bcrs-shared-components/folio-number-input",
-    "version": "1.0.0",
-    "publishConfig": {
-        "access": "public"
-    },
-    "dependencies": {
-        "@bcrs-shared-components/interfaces": "^1.0.22",
-        "vue-property-decorator": "^8.4.0"
-    }
-}
