+++ src/components/folio-number-input/index.ts
@@ -1,0 +1,1 @@
+export { default as FolioNumberInput } from './FolioNumberInput.vue'
